requirements.txt
drf-spectacular>=0.14.0
djangorestframework>=3.14.0
python-decouple>=3.8
pytest-django>=4.5, <5.0
pytest>=7.0, <8.0
django-dbbackup>=4.0, <5.0
